[WM] BindResource refactoring and overloading
* A new BindResource() version has been introduce to allows the policy to
specificy a resource path and a ResourceBitset object to directly map the
path to a set of resource descriptors. The ResourceBitset set bits refer to
the id numbers of the descriptors to bind to.

* The reference numbering of the binding maps has been changed. The map
has been replaced by a vector. Therefore we do not need anymore of the
hash function to generate a key. This change has been introduce to make
possible the use of both the version of BindResource(), without struggling
to retrieve the on going binding map.

* The function BindingStr() became useless and it has been removed.

* The vector of scheduling binding maps is cleared once the synchronization
is performed (SetResourceBinding() call).

+++ include/bbque/app/working_mode_status.h
@@ -153,7 +153,7 @@ public:
 	 *
 	 * @return A map of Usage objects
 	 */
-	virtual br::ResourceAssignmentMapPtr_t GetSchedResourceBinding(size_t b_id = 0) const = 0;
+	virtual br::ResourceAssignmentMapPtr_t GetSchedResourceBinding(uint32_t b_id = 0) const = 0;
 
 	/**
 	 * @brief Get the bitmap of the clusters currently used.
@@ -163,8 +163,7 @@ public:
 	 *
 	 * @return A bitset data structure
 	 */
-	virtual br::ResourceBitset BindingSet(
-		const br::ResourceType & r_type) const = 0;
+	virtual br::ResourceBitset BindingSet(const br::ResourceType & r_type) const = 0;
 
 	/**
 	 * @brief Get the bitmap of the clusters previously used.
@@ -174,8 +173,7 @@ public:
 	 *
 	 * @return A bitset data structure
 	 */
-	virtual br::ResourceBitset BindingSetPrev(
-		const br::ResourceType & r_type) const = 0;
+	virtual br::ResourceBitset BindingSetPrev(const br::ResourceType & r_type) const = 0;
 
 	/**
 	 * @brief Check if clusters used are changed
